Throw exceptions on errors, check number of arguments, fix in tdControllerValue
